Split site-details-page.tsx into feature-scoped components
The 1355-line file mixed 5 tabs, 5 dialogs, and Git operations in one
component. Split into sibling files under src/features/sites/ (no new
subdirectory - matches how backups/database/logs/snapshots already
keep their pages at the feature-folder root), following the same
prop-threading precedent as project-wizard.tsx: state stays flat in
the orchestrator, extracted pieces receive values/setters as props.

The one exception is useSiteGit: gitStatus/gitDetails are read from
4 separate JSX locations (Overview summary, Tools-tab Git card, the
delete-dirty-check, and the checkout-conflict dialog), and 3 of 4 git
handlers touch the shared busy/message mutex - a hook can fan state
out to sibling subtrees where a plain component can't.

No behavior change intended - same tabs, same dialogs, same Tauri
commands. main.tsx is untouched since the orchestrator kept its path.
